feat: add theme to all stories
